% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bass.R
\name{bass}
\alias{bass}
\title{Compute Bass slope coefficient / A component}
\usage{
bass(M, p, q)
}
\arguments{
\item{M}{Current market share, numeric 0-1}

\item{p}{Innovation parameter, numeric 0-1}

\item{q}{Immitation parameter, numeric 0-1}
}
\value{
Slope of bass curve with paramers p, q at market penetration M (numeric)
}
\description{
Compute Bass slope coefficient given current market share and bass parameters p, q
}
\examples{
# compute slope of the curve with parameters (0.05, 0.4) when half the market has been reached:
bass(0.5, 0.05, 0.4)
}
