% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_bp.R
\name{calculate_bp}
\alias{calculate_bp}
\title{Calculate base penetration rate for arbitrary variables}
\usage{
calculate_bp(prepped_data, fixed_predictor, other_predictors, coefficients,
  id_col, frame = 10)
}
\arguments{
\item{prepped_data}{data frame formatted correctly for this model, probably from <link to setup function>}

\item{fixed_predictor}{unquoted name of explanatory variable that will have constant effect of 1 (usually income)}

\item{other_predictors}{vector of}

\item{coefficients}{named list or vector of parameters}

\item{id_col}{unquoted name of column containing observation unique IDs}

\item{frame}{numeric; maximum divergence allowed for calculated BP. By default, the smallest BP will be 0.1*mean, and the largest will be 10*mean)}
}
\description{
ToDo: fix the error handling a little, it's gross rn
}
