% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{calculate_target}
\alias{calculate_target}
\title{Calculate target number to add from target value columns}
\usage{
calculate_target(prepped_data, ...)
}
\arguments{
\item{prepped_data}{data frame prepped for use elsewhere in this model}

\item{...}{variables with}
}
\value{
list of number added in each time step
}
\description{
Documentation needs to be better here
}
