% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{center_and_scale}
\alias{center_and_scale}
\title{Center and scale a single vector}
\usage{
center_and_scale(x, na.rm = TRUE)
}
\arguments{
\item{x}{numeric, a vector containing the values to be centered and scaled}

\item{na.rm}{a logical value indicating whether NA values should be stripped before the computation proceeds.}
}
\value{
A vector of the same length as \code{x}, but with mean=0 and sd=1
}
\description{
Create a vector of values with mean 0 and sd 1, which can be useful for model estimation
}
\examples{
center_and_scale(1:3) # = c(-1, 0, 1)
center_and_scale(1:4) # = c(-1.1618950, -0.3872983,  0.3872983,  1.1618950)
}
