% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{center_and_scale_validation}
\alias{center_and_scale_validation}
\title{Run center-and-scale operation on one dataset using parameters from another}
\usage{
center_and_scale_validation(validation, training, ..., na.rm = TRUE)
}
\arguments{
\item{validation}{data frame of data that will be used}

\item{training}{data frame from which means and sds will be drawn, must have at least all cols in \code{...}}

\item{...}{variables present in both \code{validation} and \code{training}}

\item{na.rm}{a logical value indicating whether NA values should be stripped before the computation proceeds.}
}
\value{
data frame with same rows and columns as \code{validation} but with the variables in \code{...} centered and scaled according to distribution of same variables in \code{training}
}
\description{
again, dots are variables to center and scale, but they're centered and scaled according to training data chars
}
