% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{center_and_scale_vars}
\alias{center_and_scale_vars}
\title{Center and scale a selection of variables in a dataset place}
\usage{
center_and_scale_vars(data, ..., na.rm = TRUE)
}
\arguments{
\item{data}{data frame from which variables in \code{...} are selected}

\item{...}{variables in data to center and scale}

\item{na.rm}{a logical value indicating whether NA values should be stripped before the computation proceeds.}
}
\value{
data frame with same columns as \code{data} but with the variables in \code{...} centered and scaled
}
\description{
Apply \code{\link{center_and_scale}} over the variables named in \code{...}
}
\examples{
# center and scale the fixed predictor column in the minimal_data example dataset
minimal_data_predsCS <- center_and_scale_vars(minimal_data, fixed_pred)
mean(minimal_data_predsCS$fixed_pred) # = 0
sd(minimal_data_predsCS$fixed_pred)   # = 1
}
