% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{df_row_to_vec}
\alias{df_row_to_vec}
\title{Extract info from one row of a data frame as a named vector}
\usage{
df_row_to_vec(data, ..., .row_no = 1)
}
\arguments{
\item{data}{A data.frame or similar containing the row and columns you want to extract}

\item{...}{One or more unquoted expressions separated by commas that can be passed}

\item{.row_no}{integer, which row to grab from data}
}
\value{
Named vector
}
\description{
This is mainly being used for
}
\examples{
df_row_to_vec(data.frame(x=1, y=2, z=3), x, y, z, row_no = 1)
}
