% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_model.R
\name{estimate_iterbass}
\alias{estimate_iterbass}
\title{Estimate model coefficients for multiple years of additions}
\usage{
estimate_iterbass(
  prepped_data,
  id_col,
  targ_cols,
  fixed_predictor,
  other_predictors,
  starting_values = NULL
)
}
\arguments{
\item{prepped_data}{data frame formatted correctly for this model (mostly just needs a market_limit column)}

\item{id_col}{unquoted name of column containing observation unique IDs}

\item{targ_cols}{Vector of unquoted names of cols with EVs/Whatever present each year}

\item{fixed_predictor}{unquoted name of explanatory variable that will have constant effect of 1 (usually income)}

\item{other_predictors}{vector of unquoted variable names of explanatory variables for which coefficient will be estimated}

\item{starting_values}{named vector of starting values that will override default parameter starts, this can include any variable in \code{other_predictors}, 'intercept', 'p', and 'q}
}
\value{
a list containing the following elements:
\itemize{
\item \code{coeffs} a named vector of estimated coefficients
\item \code{MSE} numeric combined MSE from all years (FUTURE: will have this split out by year)
\item \code{iterations} number of iterations before convergence
\item \code{preds} named list of prediction vectors
}
}
\description{
Estimate model coefficients for multiple years of additions
}
