% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_model.R
\name{estimate_model}
\alias{estimate_model}
\title{Estimate model coefficients for multiple years of additions}
\usage{
estimate_model(prepped_data, id_col, targ_cols, n_to_add, fixed_predictor,
  other_predictors)
}
\arguments{
\item{prepped_data}{data frame formatted correctly for this model}

\item{id_col}{unquoted name of column containing observation unique IDs}

\item{targ_cols}{Vector of unquoted names of cols with EVs/Whatever present each year}

\item{n_to_add}{Vector of number of EVs/whatever to add each year year (length 1 - length of targ_cols)}

\item{fixed_predictor}{unquoted name of explanatory variable that will have constant effect of 1 (usually income)}

\item{other_predictors}{vector of unquoted variable names of explanatory variables for which coefficient will be estimated}
}
\value{
Named vector of coefficient estimates
}
\description{
Estimate model coefficients for multiple years of additions
}
