% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_bp.R
\name{grab_single_symbol}
\alias{grab_single_symbol}
\title{Try-Catch to make sure function receives only a single variable name when needed}
\usage{
grab_single_symbol(arg)
}
\arguments{
\item{arg}{an argument passed from a previous function}
}
\value{
Either variable name as text or error
}
\description{
I have no idea if this works right.
}
