% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_model.R
\name{mse_model}
\alias{mse_model}
\title{Calculate summed error over arbitrary number of years of additions}
\usage{
mse_model(
  par,
  prepped_data,
  targ_cols,
  fixed_predictor,
  other_predictors,
  id_col,
  params_order,
  return_preds = FALSE,
  frame = 10
)
}
\arguments{
\item{par}{numeric vector used by optimx}

\item{prepped_data}{data frame formatted correctly for this model}

\item{targ_cols}{Vector of unquoted names of cols with EVs/Whatever present each year}

\item{fixed_predictor}{unquoted name of explanatory variable that will have constant effect of 1 (usually income)}

\item{other_predictors}{vector of unquoted variable names of explanatory variables for which coefficient will be estimated}

\item{id_col}{unquoted name of column containing observation unique IDs}

\item{params_order}{character vector, names corresponding to par (necessary because optimx strips names)}

\item{return_preds}{logical, should this method return predictions only (stop before calculating mse)}

\item{frame}{numeric; maximum divergence allowed for calculated BP. By default, the smallest BP will be 0.1\*mean, and the largest will be 10\*mean)}
}
\value{
Combined mean squared error for a multiple assignment runs over successive years
}
\description{
This function calculates the overall mean squared error from a set of projections over various years given a set of coefficients.
Should work as long as at least two columns are included in targ_cols
}
\examples{
p_start <- c('intercept'=1, 'other_pred1'=1, 'other_pred2'=1, 'p'=0.1, 'q'=0.4)
mse_model(par = p_start,
          prepped_data = minimal_data,
          targ_cols = c(val_start:val_s2),
          fixed_predictor = fixed_pred,
          other_predictors = c(other_pred1, other_pred2),
          id_col = location,
          params_order = names(p_start))
}
