% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_model.R
\name{mse_single}
\alias{mse_single}
\title{Calculate error from a single year run}
\usage{
mse_single(prepped_data_bp, starting_vals, target_vals, n_to_add, p, q)
}
\arguments{
\item{prepped_data_bp}{data frame formatted correctly for this model, with bp column added}

\item{starting_vals}{Unquoted name of starting value column (number in year n)}

\item{target_vals}{Unquoted name of target value column (number in year n+1)}

\item{n_to_add}{Number of EVs/whatever to add this year}

\item{p}{Innovation parameter, numeric 0-1}

\item{q}{Immitation parameter, numeric 0-1}
}
\description{
This'll mostly get called from within mse_multiple
}
