% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_predictions.R
\name{predict_iterbass}
\alias{predict_iterbass}
\title{Get predictions from iterative bass model}
\usage{
predict_iterbass(
  prepped_data,
  id_col,
  start_col,
  n_to_add,
  coeffs,
  fixed_predictor,
  other_predictors,
  prediction_col = "pred",
  frame = 10
)
}
\arguments{
\item{prepped_data}{data frame formatted correctly for this model}

\item{id_col}{unquoted name of column containing observation unique IDs}

\item{start_col}{unquoted names of column used as starting value for predictions}

\item{n_to_add}{numeric, total number of things to assign across all zones}

\item{coeffs}{named vector of model parameters, must include every variable in \code{other_predictors}, as well as 'intercept', 'p', and 'q}

\item{fixed_predictor}{unquoted name of explanatory variable that will have constant effect of 1 (usually income)}

\item{other_predictors}{vector of unquoted variable names of explanatory variables for which coefficient will be estimated}

\item{prediction_col}{unquoted or quated name for column to assing predictions to, defaults to \code{pred}}

\item{frame}{numeric; maximum divergence allowed for calculated BP. By default, the smallest BP will be 0.1\*mean, and the largest will be 10\*mean)}
}
\value{
Data frame formatted as \code{prepped_data} but with an added column of predictions
}
\description{
Predicts values given a correctly formatted dataset, set of parameters, starting value, and number of items to distribute
FUTURE: if I can figure out how to get BP to work without having whole dataset as input, this will also input/output vectors
}
