% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{prep_data}
\alias{prep_data}
\title{Prepare data for stocked bass modeling}
\usage{
prep_data(
  raw_data,
  id_cols,
  targ_cols,
  predictor_cols,
  ml_col,
  bump_ml = FALSE,
  training_data = NULL
)
}
\arguments{
\item{raw_data}{data frame}

\item{id_cols}{unquoted name of column containing observation unique IDs as well as any other variables to preserve}

\item{targ_cols}{vector of unquoted names of cols with EVs/Whatever present each year}

\item{predictor_cols}{vector of unquoted variable names of explanatory variables for which coefficient will be estimated}

\item{ml_col}{unquoted name of column containing the market limit for each unit}

\item{bump_ml}{logical, should market limit be set to highest targ value if that is higher than market limit? Will throw error if any target value is higher than market limit otherwise}

\item{training_data}{training data to use, if \code{NULL} data will be prepped as training data}
}
\value{
data frame prepped for use with the rest of the functions in this package
}
\description{
This will center and scale predictor variables, add market limit, and remove everything else
}
