% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_assignment.R
\name{run_assignment2}
\alias{run_assignment2}
\title{Iterate over the model once, completing all assignment steps}
\usage{
run_assignment2(market_current, market_limit, base_rate, n_to_add, p, q,
  tot_iters = 40)
}
\arguments{
\item{market_current}{numeric, vector of current market values}

\item{market_limit}{numeric, vector of market limits}

\item{base_rate}{numeric, vector of base penetration rates}

\item{n_to_add}{numeric, total number of things to assign across all zones}

\item{p}{Innovation parameter, numeric 0-1}

\item{q}{Immitation parameter, numeric 0-1}

\item{tot_iters}{integer > 0, number of iterations to run}
}
\value{
numeric vector of
}
\description{
This functions runs all iterations for a single year step
}
