% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_assignment.R
\name{run_assignment_with_variable_bp}
\alias{run_assignment_with_variable_bp}
\title{Run the assignment with a given set of parameters}
\usage{
run_assignment_with_variable_bp(prepped_data, params, tot_evs,
  bpr_frame = 10, tot_iters = 40)
}
\arguments{
\item{prepped_data}{Data frame formatted correctly for this model, probably from <link to setup function>}

\item{params}{Named list or vector of parameters, as of right now must include all of BP_intercept, BP_coef_commute, BP_coef_gravity, BP_coef_carpool, innov_coef_p, and innov_coef_p}

\item{tot_evs}{Total number objects to assign, numeric}

\item{bpr_frame}{Maximum divergence allowed for calculated BPR. By default (at 10, the smallest BPR will be 0.1*mean, and the largest will be 10*mean)}

\item{tot_iters}{Number of iterations to run, integer > 0}
}
\value{
Data frame matching format of \code{prepped_data} with all objects assigned.
}
\description{
\code{\link[=run_assignment]{run_assignment()}} assumes BPR is fixed / precalculated, so this first computes it for a given set of parameters then runs the assignment.
This function also keeps a frame on BPR, by default an order of magnitude in either direction
}
