% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_model.R
\name{mse_multiple}
\alias{mse_multiple}
\title{Calculate summed error over arbitrary number of years of additions}
\usage{
mse_multiple(par, prepped_data, targ_cols, n_to_add, fixed_predictor,
  other_predictors, id_col, params_order, frame = 10)
}
\arguments{
\item{par}{numeric vector used by optimx}

\item{params_order}{character vector, names corresponding to par}

\item{frame}{}
}
\description{
Calculate summed error over arbitrary number of years of additions
}
