% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_assignment_step.R
\name{single_assignment_step}
\alias{single_assignment_step}
\title{Run a single assignment step for the model}
\usage{
single_assignment_step(data, n_this_iter, iteration, p, q)
}
\arguments{
\item{data}{Data frame output from previous step}

\item{n_this_iter}{Number to assign this step, probably total / n_steps}

\item{iteration}{Iteration step, integer > 0 ... pretty much ignored silently}

\item{p}{Innovation parameter, numeric 0-1}

\item{q}{Immitation parameter, numeric 0-1}
}
\value{
Data frame matching format of \code{data}.
}
\description{
Run a single round of the model, assigning units equal to total_to_assign / n_steps distributed among the regions included in the model
This should only be called by \code{\link[=run_assignment]{run_assignment()}}
}
